test(cozy-stack-client): Gather two tests and fix them
